requirements.txt
langchain
streamlit
python-dotenv
pillow
opencv-python
transformers
librosa
google-generativeai
langchain-google-genai
PyPDF2
utils